Time: 0 ms (100%), Space: 42 MB (5.36%) - LeetHub

+++ 1584-average-salary-excluding-the-minimum-and-maximum-salary/1584-average-salary-excluding-the-minimum-and-maximum-salary.java
@@ -1,13 +1,13 @@
 class Solution {
     public double average(int[] salary) {
         Arrays.sort(salary);
-        int n=salary.length;
-       int c=0;
-       int b=n-2;
+      double n=salary.length;
+      double c=0;
+      double b=n-2;
        for(int i=1;i<=n-2;i++){
         c=c+salary[i];
        }
-       double d=(double)c/b;
+       double d=c/b;
        return d;
       
     }
